requirements.txt
nicegui
fastapi
asyncpg
python-dotenv
pandas
pillow
uvicorn
